Complete the APIs for Task

+++ app/routers/user.py
@@ -1,56 +1,61 @@
 from fastapi import APIRouter, Depends, status
+from services.auth import token_interceptor
 from models.user import UserViewModel, UserCreateOrUpdateModel
 from database import get_db_context
 from sqlalchemy.orm import Session
 from services import user as user_service
 from uuid import UUID
+from schemas.user import User
 from utilities.utils import http_exception
 
 router = APIRouter(prefix="/users", tags=["User"])
 
 @router.get("", status_code=status.HTTP_200_OK)
-async def get_all_user(db: Session = Depends(get_db_context)) -> list[UserViewModel]:
+async def get_all_user(
+    user: User = Depends(token_interceptor),
+    db: Session = Depends(get_db_context)) -> list[UserViewModel]:
     """ Get all users
-
+        - If the user is admin -> Get all users in the same company
+        - If the usre is non admin -> Get the user from the token (logged user)
     Args:
+        user (User, optional): User from token. Defaults to Depends(token_interceptor).
         db (Session, optional): Db context. Defaults to Depends(get_db_context).
 
     Returns:
-        list[UserViewModel]: List of users
+        list[UserViewModel]: A list of users
     """
-    return user_service.get_all_users(db)
+    if not user.is_admin:
+        return [user_service.get_user_by_id(user.id, db)]
+    
+    return user_service.get_users_by_company_id(user.company_id, db)
 
 @router.get("/{id}", status_code=status.HTTP_200_OK)
-async def get_user_by_id(id: UUID, db: Session = Depends(get_db_context)) -> UserViewModel:
-    """ Get user by id
-
+async def get_user_by_id(
+    id: UUID,
+    user: User = Depends(token_interceptor),
+    db: Session = Depends(get_db_context)) -> UserViewModel:
+    """ Get user by Id
+        - TODO: Need to check if user id to get has the same company with the logged user
     Args:
         id (UUID): Id of the user
+        user (User, optional): User from token. Defaults to Depends(token_interceptor).
         db (Session, optional): Db context. Defaults to Depends(get_db_context).
 
     Raises:
-        http_exception: 404 error if the user could not be founs
+        http_exception: 403 Forbbiden in case the user does not have permission to do this action (not admin)
+        http_exception: 404 Not found in case the user could not be found
 
     Returns:
-        UserViewModel: A user view model
+        UserViewModel: A user model
     """
+    # Not user admin
+    if not user.is_admin and user.id != id:
+        raise http_exception(403, "You don't have permission to do this action")
     result = user_service.get_user_by_id(id, db)
     if not result:
         raise http_exception(404, "The user could not be found")
     return result
-
-@router.get("/company/{company_id}", status_code=status.HTTP_200_OK)
-async def get_users_by_company_id(company_id: UUID, db:Session = Depends(get_db_context)) -> list[UserViewModel]:
-    """ Get all users by company id
-
-    Args:
-        company_id (UUID): Company Id
-        db (Session, optional): Db context. Defaults to Depends(get_db_context).
-
-    Returns:
-        list[UserViewModel]: A list of user models
-    """
-    return user_service.get_users_by_company_id(company_id, db)
+    
 
 @router.post("", status_code=status.HTTP_201_CREATED)
 async def create_a_new_user(model: UserCreateOrUpdateModel, db: Session = Depends(get_db_context))->bool:
@@ -61,9 +66,9 @@ async def create_a_new_user(model: UserCreateOrUpdateModel, db: Session = Depend
         db (Session, optional): Db context. Defaults to Depends(get_db_context).
 
     Raises:
-        http_exception: 404 Not found
-        http_exception: 409 Conflict
-        http_exception: 500 Internal error
+        http_exception: 404 Not found in case the company could not be found
+        http_exception: 409 Conflict in case the user_name or email exist
+        http_exception: 500 Internal error for the server error
 
     Returns:
         bool: True if sucess
@@ -81,21 +86,30 @@ async def create_a_new_user(model: UserCreateOrUpdateModel, db: Session = Depend
         
 
 @router.put("/{id}", status_code=status.HTTP_200_OK)
-async def update_a_user(id: UUID, model: UserCreateOrUpdateModel, db: Session = Depends(get_db_context)) -> bool:
+async def update_a_user(
+    id: UUID,
+    model: UserCreateOrUpdateModel,
+    user: User = Depends(token_interceptor),
+    db: Session = Depends(get_db_context)) -> bool:
     """ Update a user
-
+        - TODO: Need to check if user id to update has the same company with the logged user
     Args:
-        id (UUID): User id to update
-        model (UserCreateOrUpdateModel): User model to update
+        id (UUID): Id of the user
+        model (UserCreateOrUpdateModel): user model to update
+        user (User, optional): User from token. Defaults to Depends(token_interceptor).
         db (Session, optional): Db context. Defaults to Depends(get_db_context).
 
     Raises:
-        http_exception: 404 Not found
-        http_exception: 500 Internal error
+        http_exception: 403 Forbiden in case user is non admin and try to update for other user
+        http_exception: 404 Not found in case the user could not be found to update
+        http_exception: 409 Conflict in case the user_name or email exist
+        http_exception: 500 Internal server for the server error
 
     Returns:
-        _type_: True if success
+        bool: True if success
     """
+    if not user.is_admin and user.id != id:
+        raise http_exception(403, "You have permission to do this action")
     result = user_service.create_or_update_user(db, model, id)
     match result:
         case status.HTTP_200_OK:
@@ -109,20 +123,28 @@ async def update_a_user(id: UUID, model: UserCreateOrUpdateModel, db: Session = 
         
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_a_user(id: UUID, db: Session = Depends(get_db_context)) -> None:
-    """ Delete a user by Id
+async def delete_a_user(
+    id: UUID,
+    user: User = Depends(token_interceptor),
+    db: Session = Depends(get_db_context)) -> None:
+    """ Soft delete a user
+       - TODO: Need to check if user id detele get has the same company with the logged user
 
     Args:
-        id (UUID): Id of the user
+        id (UUID): User id
+        user (User, optional): User from token. Defaults to Depends(token_interceptor).
         db (Session, optional): Db context. Defaults to Depends(get_db_context).
 
     Raises:
-        http_exception: 404 Not found
-        http_exception: 500 Internal error
+        http_exception: 403 Forbiden in case the user does not have permission to do this action (non admin)
+        http_exception: 404 Not found in case the user could not be found to delete
+        http_exception: 500 Internal error in case internal server error
 
     Returns:
-        _type_: _description_
+        _type_: 204 No content
     """
+    if not user.is_admin and user.id != id:
+        raise http_exception(403, "You don't have permission to do this action")
     result = user_service.delete_a_user(id, db)
     match result:
         case status.HTTP_404_NOT_FOUND:
